Improved memory utilization efficiency and Obsoleted pg_show_plans.enable_txid.
[1] Improved memory utilization efficiency.
In order to reduce the consumption of the shared memory, all nested level plans of a query are stored in a relevant hash entry.

[2] Obsoleted pg_show_plans.enable_txid.
The parameter pg_show_plans.enable_txid was introduced to avoid the conflict, which makes a server failure, between invoking GetTopTransactionId() and The pg_create_logical_replication_slot() function. However, GetTopTransactionId() was removed from the source code, therefore, this parameter is no longer needed.