Fix #29; Fix incorrect offset calculation
To calculate "offset" to store a new plan at, we should sum the lengths
of the already stored plans within the memory plus the number of plans
to accomodate for NULL character at the end of each plan.

Instead, I used to sum up the lengths of all plans and add 1.