Mark functions as static everywhere
Fundamentally, this is an unnecessary exercise, as the first declaration
or definition of a function determines whether it is static or not, and
the keyword "static" is not needed later on.

However, we got a complaint on #30 that gcc 4.8.5 with the option
-std=gnu99 complains about the lack of "static" in one place.  This is a
compiler bug that has since been fixed.

Still, let's add "static" everywhere.  It does not hurt and might make
life easier for the casual reader.

Per suggestion from Linn Zhang.

This closes #30.